Update documentation for rlcompleter.py
Hi,

I want to document this because this issue was very hard to google and figure out, so it will hopefully help others. I think this repo might be a reasonable place to put this, but this is a combination of multiple things. The summary is when I tab-comlete in python, it shows up as very angry flashing red, see

https://asciinema.org/a/A1wIPeDTmoSq8NOoPJQ0kgTvm

The problem is fundamentally the default python/readline tab completion highlights files, and most missing python things aren't existing files (you can test this by creating a file say `__doc__` and hitting tab).

You can also fix this with the following inputrc

```
$if python
  set colored-stats off
$endif
```

See
https://unix.stackexchange.com/questions/605166/why-does-setting-mi-in-ls-colors-effect-postgresql-and-python-tab-completion